[generator] Break out a few more classes to files.
