+++ tools/generator/CodeGenerationTarget.cs
@@ -1,0 +1,9 @@
+namespace Xamarin.Android.Binder
+{
+	public enum CodeGenerationTarget
+	{
+		XamarinAndroid,
+		XAJavaInterop1,
+		JavaInterop1,
+	}
+}
