+++ tools/generator/Java.Interop.Tools.Generator.Importers/Xml/InterfaceXmlGenBaseSupport.cs
@@ -1,0 +1,18 @@
+﻿using System.Xml.Linq;
+
+namespace MonoDroid.Generation
+{
+	public class InterfaceXmlGenBaseSupport : XmlGenBaseSupport
+	{
+		public InterfaceXmlGenBaseSupport (XElement pkg, XElement elem)
+			: base (pkg, elem)
+		{
+		}
+		
+		public override string TypeNamePrefix {
+			get { return (IsPrefixableName (RawName) ? "I" : string.Empty); }
+		}
+	}
+}
+
+
